refactor: split PredictionChart into focused sub-components
